✨ Normalize previousSimulation code
