+++ mon-entreprise/source/reducers/rootReducer.ts
@@ -2,13 +2,13 @@ import { Action } from 'Actions/actions'
 import { defaultTo, omit, without } from 'ramda'
 import reduceReducers from 'reduce-reducers'
 import { combineReducers, Reducer } from 'redux'
-import { SavedSimulation } from 'Selectors/storageSelectors'
+import { PreviousSimulation } from 'Selectors/previousSimulationSelectors'
 import { DottedName } from 'modele-social'
 import { objectifsSelector } from '../selectors/simulationSelectors'
 import inFranceAppReducer, { Company } from './inFranceAppReducer'
-import storageRootReducer from './storageReducer'
 import { Names } from 'modele-social/dist/names'
 import { getCompanySituation } from 'Components/utils/useSimulationConfig'
+import previousSimulationRootReducer from './previousSimulationRootReducer'
 
 function explainedVariable(
 	state: DottedName | null = null,
@@ -165,15 +165,15 @@ const existingCompanyReducer = (state: RootState, action: Action) => {
 const mainReducer = combineReducers({
 	explainedVariable,
 	simulation,
-	previousSimulation: defaultTo(null) as Reducer<SavedSimulation | null>,
+	previousSimulation: defaultTo(null) as Reducer<PreviousSimulation | null>,
 	activeTargetInput,
 	inFranceApp: inFranceAppReducer,
 })
 
 export default reduceReducers<RootState>(
-	mainReducer,
+	mainReducer as any,
 	existingCompanyReducer as Reducer<RootState>,
-	storageRootReducer as Reducer<RootState>
+	previousSimulationRootReducer as Reducer<RootState>
 ) as Reducer<RootState>
 
 export type RootState = ReturnType<typeof mainReducer>
